pyproject.toml
[tool.poetry]
name = "estudo-airflow"
version = "0.1.0"
description = ""
authors = ["leonardo <leonardoaugusto1000a@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
apache-airflow = "^3.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
